Separate the ModuleInstanceType class
